requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
scipy>=1.7.0
tqdm>=4.60.0
pillow>=8.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
einops>=0.6.0
geoopt>=0.5.0
gsplat>=0.1.0
torch-geometric>=2.3.0
timm>=0.9.0